Avoid bare 'except:' in geo_dist.py
So-called 'bare' excepts, i.e. where the type of the exception is not defined,
will catch things one might not want to catch, such as Python-raised errors like SystemExit
or KeyboardInterrupt exceptions triggered by the user.

Except-clauses are used to catch _known_ failure states. For unknown and unexpected error
states, we want the program to crash